feat: move get_region_name to services
